users & games crud operations

+++ routes/gameRoutes.js
@@ -1,17 +1,31 @@
 const express = require("express");
 const { query } = require("../config/db");
 const Game = require("../models/Game");
-const { validateGame } = require("../middleware/validation");
 const { authenticate } = require("../middleware/auth");
+const { placeholder } = require("../middleware/placeholder");
 const router = express.Router();
 
 // Create
-router.post("/create", async (req, res, next) => {
-  const { title, description, htmlCode, cssCode, jsCode, authorId } = req.body;
+router.post("/create", authenticate, placeholder, async (req, res, next) => {
+  const userId = req?.user?.id;
+
+  if (!userId) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
+
+  const {
+    title,
+    description = "",
+    htmlCode = "",
+    cssCode = "",
+    jsCode = "",
+    published = false,
+  } = req.body;
+
   try {
     const result = await query(
-      "INSERT INTO games (title, description, html_code, css_code, js_code, author_id) VALUES ($1, $2, $3, $4, $5, $6) RETURNING *",
-      [title, description, htmlCode, cssCode, jsCode, authorId]
+      "INSERT INTO games (title, description, html_code, css_code, js_code, published, user_id) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *",
+      [title, description, htmlCode, cssCode, jsCode, published, userId]
     );
     const createdGame = new Game(
       result.rows[0].id,
@@ -20,7 +34,7 @@ router.post("/create", async (req, res, next) => {
       result.rows[0].html_code,
       result.rows[0].css_code,
       result.rows[0].js_code,
-      result.rows[0].author_id
+      result.rows[0].user_id
     );
     res.status(201).json(createdGame);
   } catch (error) {
@@ -69,13 +83,25 @@ router.get("/user/:userId", async (req, res, next) => {
 });
 
 // Update
-router.put("/update/:id", async (req, res, next) => {
+router.put("/update/:id", authenticate, async (req, res, next) => {
   const { id } = req.params;
-  const { title, description, htmlCode, cssCode, jsCode } = req.body;
+  const game = await Game.findById(id);
+
+  if (!game) {
+    return res.status(404).json({ message: `Game with ID: ${id} not found` });
+  }
+
+  const { title, description, htmlCode, cssCode, jsCode, published } = req.body;
+  const userId = req?.user?.id;
+
+  if (!userId || userId.toString() !== game.user_id.toString()) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
+
   try {
     await query(
-      "UPDATE games SET title = $1, description = $2, html_code = $3, css_code = $4, js_code = $5 WHERE id = $6",
-      [title, description, htmlCode, cssCode, jsCode, id]
+      "UPDATE games SET title = $1, description = $2, html_code = $3, css_code = $4, js_code = $5, published = $6 WHERE id = $7",
+      [title, description, htmlCode, cssCode, jsCode, published, id]
     );
     res.status(200).json({ message: "Game updated" });
   } catch (error) {
@@ -84,45 +110,46 @@ router.put("/update/:id", async (req, res, next) => {
 });
 
 // Delete
-router.delete(
-  "/delete/:id",
-  // authenticate,
-  // validateGame,
-  async (req, res, next) => {
-    const { id } = req.params;
-    try {
-      await query("DELETE FROM games WHERE id = $1", [id]);
-      res.status(200).json({ message: "Game deleted" });
-    } catch (error) {
-      next(error);
-    }
+router.delete("/delete/:id", authenticate, async (req, res, next) => {
+  const { id } = req.params;
+  const userId = req?.user?.id;
+
+  if (!userId) {
+    return res.status(401).json({ message: "Unauthorized" });
   }
-);
 
-router.post("/saveToGitHub", async (req, res) => {
   try {
-    const { gameId } = req.body;
-    const game = await getGameById(gameId); // Assuming you have this function
-    if (!game) {
-      return res.status(404).json({ message: "Game not found" });
-    }
-
-    const github = new Octokit({ auth: req.user.githubToken }); // Assuming GitHub token is stored in req.user
-    const gist = await github.gists.create({
-      files: {
-        "index.html": { content: game.htmlCode },
-        "styles.css": { content: game.cssCode },
-        "script.js": { content: game.jsCode },
-      },
-      description: game.name,
-      public: true,
-    });
-
-    return res.json({ gistUrl: gist.data.html_url });
+    await query("DELETE FROM games WHERE id = $1", [id]);
+    res.status(200).json({ message: "Game deleted" });
   } catch (error) {
-    console.error(error);
-    res.status(500).json({ message: "An error occurred" });
+    next(error);
   }
 });
+
+// router.post("/saveToGitHub", async (req, res) => {
+//   try {
+//     const { gameId } = req.body;
+//     const game = await getGameById(gameId); // Assuming you have this function
+//     if (!game) {
+//       return res.status(404).json({ message: "Game not found" });
+//     }
+
+//     const github = new Octokit({ auth: req.user.githubToken }); // Assuming GitHub token is stored in req.user
+//     const gist = await github.gists.create({
+//       files: {
+//         "index.html": { content: game.htmlCode },
+//         "styles.css": { content: game.cssCode },
+//         "script.js": { content: game.jsCode },
+//       },
+//       description: game.name,
+//       public: true,
+//     });
+
+//     return res.json({ gistUrl: gist.data.html_url });
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({ message: "An error occurred" });
+//   }
+// });
 
 module.exports = router;
